+ req check pedidos / fechaBaja

+++ src/main/java/ms/usuario/service/impl/ClienteServiceImpl.java
@@ -1,9 +1,10 @@
 package ms.usuario.service.impl;
 
+import java.time.LocalDate;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,7 @@ import ms.usuario.dao.ClienteRepository;
 import ms.usuario.domain.Cliente;
 import ms.usuario.exceptions.RiesgoException;
 import ms.usuario.service.ClienteService;
+import ms.usuario.service.PedidoService;
 import ms.usuario.service.RiesgoCrediticioService;
 
 @Service
@@ -23,28 +25,62 @@ public class ClienteServiceImpl implements ClienteService{
 	@Autowired
 	RiesgoCrediticioService riesgoService;
 
+	@Autowired
+	PedidoService pedidoService;
+
 	@Override
 	public Optional<Cliente> buscarPorId(Integer id) {
-		return this. clienteRepo.findById(id);
+		Optional<Cliente> cliente = this.clienteRepo.findById(id);
+
+		if(cliente.isPresent() && cliente.get().getFechaBaja() == null)
+			return cliente;
+
+		return Optional.empty();
 	}
+
 	@Override
 	public Optional<Cliente> findByCuit(String cuit) {
-		return clienteRepo.findByCuit(cuit);
+		Optional<Cliente> cliente = this.clienteRepo.findByCuit(cuit);
+
+		if(cliente.isPresent() && cliente.get().getFechaBaja() == null)
+			return cliente;
+
+		return Optional.empty();
 	}
+
 	@Override
 	public Optional<Cliente> findByRazonSocial(Optional<String> razonSocial) {
-		return clienteRepo.findByRazonSocial(razonSocial);
+		Optional<Cliente> cliente = this.clienteRepo.findByRazonSocial(razonSocial);
+
+		if(cliente.isPresent() && cliente.get().getFechaBaja() == null)
+			return cliente;
+
+		return Optional.empty();
 	}
+
 	@Override
 	public List<Cliente> findAll() {
-		return clienteRepo.findAll();
+		return this.clienteRepo.findAll().stream()
+				.filter(unCli -> unCli.getFechaBaja() == null)
+				.collect(Collectors.toList());
+
 	}
 	@Override
-	public Cliente save(Cliente cliente) throws RiesgoException {
-		if(riesgoService.situacionBCRA(cliente.getCuit()) > 2) {
+	public Cliente save(Cliente nuevo) throws RiesgoException {
+
+		if(nuevo.getId() != null) {
+			Optional<Cliente> cliente = clienteRepo.findById(nuevo.getId());
+
+			if(cliente.isPresent()) 
+				return this.clienteRepo.save(nuevo);
+			else
+				throw new RuntimeException("Cliente no encontrado");
+		}
+		else if(riesgoService.situacionBCRA(nuevo.getCuit()) > 2) {
 			throw new RiesgoException("BCRA");
 		}
-		return clienteRepo.save(cliente);
+		else
+			return this.clienteRepo.save(nuevo);
 	}
 	@Override
 	public void update(Cliente clienteDb, Cliente nuevo) {
@@ -55,11 +91,24 @@ public class ClienteServiceImpl implements ClienteService{
 		clienteDb.setObras(nuevo.getObras());
 		clienteDb.setRazonSocial(nuevo.getRazonSocial());
 		clienteDb.setUser(nuevo.getUser());
-		clienteRepo.save(clienteDb);
+		clienteDb.setFechaBaja(nuevo.getFechaBaja());
+		this.clienteRepo.save(clienteDb);
 
 	}
 	@Override
 	public void delete(Integer id) {
-		clienteRepo.deleteById(id);
+		if(pedidoService.buscarPedidos(null,id).isEmpty())
+			clienteRepo.deleteById(id);
+		else {
+			Optional<Cliente> cliente = clienteRepo.findById(id);
+
+			if(cliente.isPresent()) {
+				//Dar de baja en 30 días
+				LocalDate fecha = LocalDate.now();
+				fecha.plusDays(30);
+				cliente.get().setFechaBaja(fecha);
+				clienteRepo.save(cliente.get());
+			}
+		}
 	}
 }
